requirements.txt
requests
pytest
transformers
torch
sentence-transformers
python-dotenv
